+++ include/ust-comm.h
@@ -132,8 +132,8 @@ struct ustcomm_ust_msg {
 		struct lttng_ust_tracer_version version;
 		struct lttng_ust_tracepoint_iter tracepoint;
 		struct {
-			uint16_t data_size;	/* following filter data */
-			uint16_t reloc_offset;
+			uint32_t data_size;	/* following filter data */
+			uint32_t reloc_offset;
 		} filter;
 	} u;
 };
